chore: refactor code for clarity (WIP)
